modularised resolver and typedefs
