+++ package.json
@@ -20,6 +20,7 @@
     "joi": "17.13.3",
     "jotai": "2.12.3",
     "labelcontainer": "0.0.36",
+    "lodash": "^4.17.21",
     "mongodb": "^5.1.0",
     "mongoose": "^8.14.1",
     "next": "15.3.1",
@@ -42,6 +43,7 @@
     "@testing-library/user-event": "^14.6.1",
     "@types/graphql": "^14.2.3",
     "@types/jest": "^29.5.14",
+    "@types/lodash": "^4.17.16",
     "@types/node": "^18.11.4",
     "@types/react-dom": "^19.0.0",
     "@typescript-eslint/parser": "^8.31.1",
