+++ src/app/api/graphql/handler.ts
@@ -1,8 +1,12 @@
 import { ApolloServer } from '@apollo/server';
-import typeDefs from './schema';
-import resolvers from './resolvers';
+import { merge } from 'lodash';
+import { typeDef as certificateType } from '../../../models/cetificate/schema';
+import { resolver as certificateResolver } from '../../../models/cetificate/resolver';
 import { startServerAndCreateNextHandler } from '@as-integrations/next';
 import { NextRequest } from 'next/server';
+
+const typeDefs = [certificateType];
+const resolvers = merge(certificateResolver);
 
 const server = new ApolloServer({
     typeDefs,
